chore: user에 받을 정보들 추가
